scripts/requirements.txt
# Requirements for Arabic Learning Scripts
# Install with: pip install -r requirements.txt

# LLM API (Gemini 2026)
google-genai>=1.0.0

# Data Processing
pandas>=2.0.0

# Progress Bar
tqdm>=4.65.0
